functions/requirements.txt
firebase_admin==6.5.0
firebase-functions==0.3.0
Pillow==10.3.0
requests==2.31.0
beautifulsoup4==4.11.1